django/akshaya_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
PyJWT==2.8.0
python-decouple==3.8
drf-yasg==1.21.7
psycopg2-binary==2.9.7
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
googletrans==4.0.0rc1
python-dotenv==1.0.0
whitenoise==6.5.0
gunicorn==21.2.0
celery==5.3.1
redis==5.0.1
django-filter==23.3